added getUserOrders()
