+++ src/controllers/cartcontroller.js
@@ -1,5 +1,4 @@
 const ShoppingCart = require('../models/shoppingcart');
-const Order = require('../models/order');
 const Response = require('../api/response');
 
 module.exports = class CartController {
@@ -109,7 +108,7 @@ module.exports = class CartController {
                                 return;
                             }
                             let success = !!updatedItem;
-                            let message = updatedItem ? `Item with productId ${productId} and quantity ${quantity} was removed from the cart` :
+                            let message = success ? `Item with productId ${productId} and quantity ${quantity} was removed from the cart` :
                                 'Item could not be removed from cart';
 
                             res.send(Response.makeResponse(success, message, updatedItem));
@@ -143,90 +142,6 @@ module.exports = class CartController {
             let userId = req.params.id;
             let quantity = req.body.quantity;
 
-
-        } catch (e) {
-            res.send(Response.makeResponse(false, e.toString()));
-        }
-    }
-
-    static createOrder(req, res) {
-        try {
-            const Product = require('../models/product.js');
-            const User = require('../models/user.js');
-            const order = new Order;
-            let jsonOrderSummary = [];
-
-            order.shippingAddress = req.body.shippingAddress;
-
-            // Every item will be bought by the same user, so find the user before the loop:
-            User.fromId(req.params.id, (err, user) => {
-                if (err) {
-                    res.send(Response.makeResponse(false, `User with id ${req.params.id} does not exist!`));
-                    return;
-                }
-                order.buyerId = user.id;
-
-                ShoppingCart.getCartItems(order.buyerId, (err, items) => {
-                    if (err || items.length === 0) {
-                        res.send(Response.makeResponse(false, `Could not get cart items`));
-                        return;
-                    }
-
-                    items.forEach((item, index, array) => {
-
-                        // Find the product:
-                        Product.fromId(item.productId, (err, prod) => {
-
-                            if (err) {
-                                res.send(Response.makeResponse(false, err.toString()));
-                                return;
-                            }
-                            // Have to recheck that the item did not run out of stock while it was sitting in the user's cart
-                            if (prod.quantity < item.quantity) {
-                                let message = `Check product "quantity": The quantity you are trying to order 
-                                    (${item.quantity}) is greater than the product stock (${prod.quantity})`;
-                                res.send(Response.makeResponse(false, message, prod));
-                                return;
-                            }
-
-                            order.sellerId = prod.sellerId;
-                            order.productId = prod.id;
-                            order.quantity = item.quantity;
-                            order.totalCost = item.quantity * prod.price;
-                            order.delivered = 0;
-
-                            order.save((err, submittedOrder) => {
-
-                                if (err) {
-                                    res.send(Response.makeResponse(false, `Error: could not place order`));
-                                    return;
-                                }
-                                jsonOrderSummary.push(submittedOrder.toJson());
-
-                                item.delete((err, removedItem) => {
-                                    if (err) {
-                                        res.send(Response.makeResponse(false, err.toString()));
-                                        return;
-                                    }
-
-                                    prod.quantity = item.quantity - order.quantity;
-                                    prod.save((err, updatedProduct) => {
-                                        if (err) {
-                                            res.send(Response.makeResponse(false, `Error: product quantity could not be updated`));
-                                            return;
-                                        }
-
-                                        if (index === array.length - 1) {
-                                            console.log('The last iteration!');
-                                            res.send(Response.makeResponse(true, "Order placed", jsonOrderSummary));
-                                        }
-                                    }, true);
-                                });
-                            });
-                        });
-                    });
-                });
-            });
 
         } catch (e) {
             res.send(Response.makeResponse(false, e.toString()));
